requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
python-dotenv
tensorflow
numpy
Pillow
twilio
geopy
email-validator